move In to the same package as InPipe & InProcessor
